### 2022-07-16 #4572 @kyleniemeyer
Hi @minaskar, unfortunately our current editors with expertise in astronomy, cosmology, and astrophysics do not have any availability to take on new submissions—you might have seen that we actually have a [call for new editors](https://blog.joss.theoj.org/2022/07/call-for-editors-astro) particularly in these areas. So, we have to waitlist your submission for the time being, until an editor becomes available.
