### 2022-07-17 89a5658
Replace os with pathlib
