### 2022-07-22 dfa07ce
attempt to fix python 3.7 state error
