+++ docs/build/html/_sources/index.rst.txt
@@ -59,17 +59,17 @@ Attribution & Citation
 Please cite the following if you find this code useful in your
 research. The BibTeX entries for the papers are::
 
-    @article{karamanis2022pmc,
+    @article{karamanis2022accelerating,
         title={Accelerating astronomical and cosmological inference with Preconditioned Monte Carlo},
-        author={Karamanis, Minas and Beutler, Florian and Peacock, John A and Nabergoj, David, and Seljak, Uro\v{s}},
-        journal={in prep},
+        author={Karamanis, Minas and Beutler, Florian and Peacock, John A and Nabergoj, David and Seljak, Uros},
+        journal={arXiv preprint arXiv:2207.05652},
         year={2022}
     }
 
     @article{karamanis2022pocomc,
         title={pocoMC: A Python package for accelerated Bayesian inference in astronomy and cosmology},
-        author={Karamanis, Minas and Nabergoj, David, and Beutler, Florian and Peacock, John A and Seljak, Uro\v{s}},
-        journal={in prep},
+        author={Karamanis, Minas and Nabergoj, David and Beutler, Florian and Peacock, John A and Seljak, Uros},
+        journal={arXiv preprint arXiv:2207.05660},
         year={2022}
     }
 
@@ -84,6 +84,10 @@ Copyright 2022 Minas Karamanis and contributors.
 
 Changelog
 =========
+
+**0.1.11 (12/07/22)**
+
+- Include saving and loading the state of the sampler. Useful for resuming runs from files.
 
 **0.1.0 (12/07/22)**
 
