### 2022-07-24 ec4ae47
add probit transform+ clip for bounded parameters
