### 2022-07-27 e36266c
bridge sampling estimator of logZ
