### 2022-08-01 #4634 @editorialbot
[REVIEW]: pocoMC: A Python package for accelerated Bayesian inference in astronomy and cosmology
**Submitting author:** @minaskar (<a href="http://orcid.org/0000-0001-9489-4612">Minas Karamanis</a>)
**Repository:** https://github.com/minaskar/pocomc
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.4
