### 2022-08-17 a1686a9
Do not evaluate the likelihood if prior is infinite
