+++ pocomc/mcmc.py
@@ -120,8 +120,10 @@ def preconditioned_metropolis(state_dict: dict,
         J_prime = numpy_to_torch(J_prime)
 
         # Compute log-likelihood, log-prior, and log-posterior
-        L_prime = numpy_to_torch(log_like(torch_to_numpy(x_prime)))
         P_prime = numpy_to_torch(log_prior(torch_to_numpy(x_prime)))
+        finite_prior_mask = torch.isfinite(P_prime)
+        L_prime = torch.full((len(x_prime),), -torch.inf)
+        L_prime[finite_prior_mask] = numpy_to_torch(log_like(torch_to_numpy(x_prime[finite_prior_mask])))
         Z_prime = numpy_to_torch(log_prob(torch_to_numpy(L_prime), torch_to_numpy(P_prime), beta))
 
         # Compute Metropolis factors
@@ -250,8 +252,10 @@ def metropolis(state_dict: dict,
         x_prime = scaler.apply_boundary_conditions(x_prime)
 
         # Compute log-likelihood, log-prior, and log-posterior
-        L_prime = log_like(x_prime)
         P_prime = log_prior(x_prime)
+        finite_prior_mask = np.isfinite(P_prime)
+        L_prime = np.full((len(x_prime),), -np.inf)
+        L_prime[finite_prior_mask] = log_like(x_prime[finite_prior_mask])
         Z_prime = log_prob(L_prime, P_prime, beta)
 
         # Compute Metropolis factor
