### 2022-08-18 2da3a3c
Fix likelihood call counts
