### 2022-08-20 minaskar/pocomc#11 @marylou-gabrie
Minor naming error in quickstart.ipynb
Two namings are used for `n_particles`/`nparticles`and `n_dim`/`ndim` in `docs/build/jupyter_execute/pages/quickstart.ipynb`.

#openjournals/joss-reviews#4634
