### 2022-08-20 66ef231
Simplify log_abs_det computation
