### 2022-08-20 minaskar/pocomc#11 @davidnabergoj
Thanks for letting us know! We've fixed this in d3698226e38ad0475314a3dddf9eb4de179f432b and the change is now visible on the [documentation page](https://pocomc.readthedocs.io/en/latest/pages/quickstart.html).
