### 2022-08-23 eeb3a95
fix bridge sampling and call counter
