### 2022-08-25 minaskar/pocomc#15 @kazewong
Typo in Readme.md example
In Basic example section of the readme, nwalkers is used in generating the prior but it is not defined. Please change it to n_particles.

#https://github.com/openjournals/joss-reviews/issues/4634
