### 2022-08-25 minaskar/pocomc#16 @kazewong
Warning in running pytest
I got these two minor warnings when I am running pytest. Please consider adjusting the testing scripts and code to remove the warnings.

```========================================================= test session starts ==========================================================
rootdir: /mnt/home/wwong/Software/PublicCode/pocomc
collected 33 items

tests/test_flow.py ..............                                                                                                [ 42%]
tests/test_pearson.py .                                                                                                          [ 45%]
tests/test_plotting.py ...                                                                                                       [ 54%]
tests/test_sampler.py .....                                                                                                      [ 69%]
tests/test_scaler.py .....                                                                                                       [ 84%]
tests/test_state.py ....                                                                                                         [ 96%]
tests/test_tools.py .                                                                                                            [100%]

=========================================================== warnings summary ===========================================================
pocomc/scaler.py:1
  /mnt/home/wwong/Software/PublicCode/pocomc/pocomc/scaler.py:1: DeprecationWarning: the imp module is deprecated in favour of importlib; see the module's documentation for alternative uses
    import imp

tests/test_flow.py::FlowTestCase::test_logprob_float64
  /mnt/home/wwong/Software/PublicCode/pocomc/pocomc/tools.py:239: UserWarning: Float64 data is currently unsupported, casting to Float32. Output will also have type Float32.
    warnings.warn(f"Float64 data is currently unsupported, casting to Float32. Output will also have type Float32.")

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
=================================================== 33 passed, 2 warnings in 58.36s
====================================================
```

#https://github.com/openjournals/joss-reviews/issues/4634
