### 2022-08-26 minaskar/pocomc#17 @kazewong
Re: Summary checkpoint in Joss review
The authors claim

1. pocoMC can speed up the sampling procedure by orders of magnitude when compared to alternatives such as nested sampling.
2. Parallelisation to computing clusters manifests linear scaling.

Regarding 1., although I do expect a speed-up coming from the reparametrization using normalizing flow, it would be better to see a demonstration of the speed-up to some degree. I know wall time demonstration is a bit annoying due to many other factors in optimizing a program, so a reduction in the total number of calls of the likelihood + benchmarking of normalizing flow would be apperciated.

Regarding 2., I don't see this parallelisation benchmark anywhere in the doc. Can the author provide a small example to show the scaling relation to some moderate number of cores (e.g. 32-128)?

#https://github.com/openjournals/joss-reviews/issues/4634
