### 2022-08-26 minaskar/pocomc#15 @kazewong
In https://pocomc.readthedocs.io/en/latest/advanced.html#parallelisation, after

```
import os

os.environ["OMP_NUM_THREADS"] = "1"
```

I think it should be beginning instead beggining.
