### 2022-08-27 7b8fec7
Fix example in README.md
