+++ docs/source/advanced.rst
@@ -17,7 +17,7 @@ If you are not familiar with these terms I encourage you to visit the :doc:`back
 
 Log-likelihood function
 -----------------------
-Suppose that we want our *likelihood* function to be a *Gaussian density* with 10 paramerers or in 10-D, we would do 
+Suppose that we want our *likelihood* function to be a *Gaussian density* with 10 parameters or in 10-D, we would do
 something like::
 
     import numpy as np
@@ -59,7 +59,7 @@ e.g. :math:`x\sim\mathcal{N}(0,3^{2})`, we would do something like::
     def log_prior(x):
         return - 0.5 * np.dot(x, x) / 3.0**2.0
 
-Alternatively, we can have priors in which not all of the paramerers have the same density a priori. For instance, suppose that 
+Alternatively, we can have priors in which not all of the parameters have the same density a priori. For instance, suppose that
 we want the first five parameters to have a flat/uniform prior :math:`x_{i}\sim\mathcal{U}(-10,10)` for :math:`i=0,1,\dots,4` and
 the last five to have a Gaussian/normal prior  :math:`x_{i}\sim\mathcal{N}(0,3^{2})` for :math:`i=5,6,\dots,9`, we would do::
 
@@ -132,7 +132,7 @@ The sampler also accepts other arguments, for a full list see :doc:`api`. Those 
   or to the log-prior using the arguments ``log_prior_args`` and ``log_prior_kwargs``.
 - The arguments ``vectorize_likelihood`` and ``vectorize_prior`` which accept boolean values allow the user to use vectorized
   log-likelihood and log-prior functions.
-- The ``periodic`` and  ``reflective`` arguments that accept a list of indeces corresponding to parameters of the model that
+- The ``periodic`` and  ``reflective`` arguments that accept a list of indices corresponding to parameters of the model that
   have *periodic* or *reflective* boundary conditions. The first kind include *phase* parameters that might be periodic e.g. 
   on a range :math:`[0,2\pi]`. The latter can arise in cases where parameters are ratios where :math:`a/b` and :math:`b/a`
   are equivalent.
@@ -151,7 +151,7 @@ The sampler also accepts other arguments, for a full list see :doc:`api`. Those 
 
     train_config = dict(validation_split = 0.2, # Percentage of particles to use for validation
                         epochs = 1000, # Maximum number of epochs
-                        batch_size = nparticles, # Batch size used for training
+                        batch_size = n_particles, # Batch size used for training
                         patience = 30, # Number of iterations to wait with no improvement in the (monitor) loss until stopping.
                         monitor = 'val_loss', # Which loss to monitor for early stopping. Options are 'val_loss' and 'loss'.
                         shuffle = True, # Shuffle the particles
@@ -196,7 +196,7 @@ works well for most target distributions.
 Correlation coefficient threshold :math:`\gamma` and number of MCMC steps
 """""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""""
 
-The :math:`\gamma` factor ( default is ``gamma = 0.75``) determnines the correlation coefficient threshold of the MCMC trajectories
+The :math:`\gamma` factor ( default is ``gamma = 0.75``) determines the correlation coefficient threshold of the MCMC trajectories
 of the particles. The lower this threshold, the more MCMC steps are performed per iteration and thus less correlated the final positions
 of the particles are with respect to their initial ones. In other words, the value :math:`\gamma` determines the number of MCMC steps
 to perform. The default value works well for most problems. If you find that your particles are not decorrelated enough and you want
@@ -224,7 +224,7 @@ This is a dictionary which includes the following arrays:
 2. ``results['loglikelihood']`` - Array with the **values of the log-likelihood** for the posterior samples given by ``results['samples']``.
[…]
