### 2022-08-27 cb214f4
Catch floating point user warning with pytest
