+++ pocomc/scaler.py
@@ -1,4 +1,3 @@
-import imp
 from typing import Union, List
 
 import numpy as np
