### 2022-08-27 579dc8e
Replace pytest with unittest to catch warnings
