### 2022-08-27 minaskar/pocomc#15 @davidnabergoj
Thanks for the issue, we have fixed the readme example in 7b8fec74801fdc2c87175859ebfbd781f40a745a and the typo in 36624396097f8f21f6a35466b7888490fa4ac395.
