### 2022-08-29 #4634 @marylou-gabrie
# A few questions:

- What does the `scale` quantity logged in the results refer to? I did not find an explanation in the documentation.

- Is there a way to access the trained flow (or flows)? This is definitely for advanced usage, but maybe it would be good to have a way to assess the quality of training of the flow model. Then, one could assess whether it is necessary to play with the training/config parameters to speedup the sampling.

- What is the training loss function for the flow? I might have missed something but I did not find this info, neither in the JOSS submission nor in the accompanying paper.

# Overall comment:
Overall the documentation is well detailed and the effort to automate the sequential scheme is remarkable. The paper gives a concise overview of the method, directed to a wise audience. I do note that the specifics of the algorithms and many related references are instead in the accompanying paper (submitted to a different venue I imagine).

# Minor:
- The correlation coefficient threshold changes notation between the "Background"section of the docs (CC) and the "Advanced Guide"(gamma).
