### 2022-08-30 #4634 @kazewong
# Questions/comments related to unchecked items

- [ ] Performance/Summary: See this thread https://github.com/minaskar/pocomc/issues/17#issue-1352355283
- [ ] Community Guideline: I don't see any comments for seeking support. Per the checklist, I think the Authors should add some guideline to seeking support.
- [ ] State-of-the-field/references: I think there are missing references which @marylou-gabrie have also flagged. For state-of-the-field, specifically related to code, I don't think it is any big issue, but the authors should add a couple references related to previous studies that didn't come with open-source codes. Once the authors added the reference suggested by @marylou-gabrie, I am happy to check both boxes out.
