### 2022-09-07 minaskar/pocomc#17 @minaskar
Thank you for your comments.

Regarding the first comment, performance claims were the subject of the accompanying paper. In that paper, we quote the number of likelihood calls for a number of artificial and realistic posterior targets. For the artificial examples, the number of calls is a good metric of the speed-up as the flow training time is negligible. However, we can provide the walltime or each example and method (PMC/SMC/NS) if you deem it necessary.

As for the second comment, we can provide a toy example in which the likelihood calculation is artificially delayed (e.g. `time.sleep(...)`) and show the speed-up for a varying number of cores.
