### 2022-09-07 #4634 @minaskar
Thank you both for your comments!

@marylou-gabrie

- The `scale` variable refers to the ratio of the Metropolis proposal scale (in latent space) to the optimal proposal scale 2.38/Sqrt(D). It is a metric of the preconditioner's performance (i.e. 1 corresponds to perfect preconditioning). I'll make sure we include an explanation in the docs.
- The trained flow can be accessed using the `Flow` class method of the sampler.
- The flow loss function is the usual forward KL divergence $D_{KL}[p|q]$.
- Cases in which PMC (and thus `pocoMC`) might not be the best choice are discussed in the Discussion section of the accompanying paper and include high-dimensional targets and/or cases in which the likelihood evaluation is cheap compared to the flow training cost. Of course, one can construct artificial cases in which the multimodality is so strong that the flow is not expressive enough to capture it. Since the code is intended for astronomical/cosmological research we do not expect any such difficulties.

@kazewong

- Regarding performance please see the response and our questions in the [aforementioned thread](https://github.com/minaskar/pocomc/issues/17#issue-1352355283).
- Brief instructions regarding users seeking support are available in the [CONTRIBUTING.md](https://github.com/minaskar/pocomc/blob/main/CONTRIBUTING.md) guide and the [main page of the docs](https://pocomc.readthedocs.io/en/latest/).
- We are more than happy to add the missing references.
