### 2022-09-17 #4634 @marylou-gabrie
Thanks @minaskar for your helpful answers.

My only remaining question is how do you evaluate the forward KL: using which samples (and potential reweighting) to approximate the expectation over $p$ (where $p$ is the succesively annealed distribution if I am not mistaken)?

Please let us know when the missing references have been added to the paper.

After that, all good for me.
