### 2022-09-19 minaskar/pocomc#17 @kazewong
@minaskar I wasn't aware of the companion paper since it is "In prep" in the draft, but I did find it on arXiv now, so I am happy with the number of call reductions stated in that paper.

For the second comment, I am not exactly sure why you would need a delayed likelihood to show this. I think it is sufficient to take either the Gaussian mixture or the Rosenbrock example, run them on a varying number of cores, record the wall time and plot the wall time against the number of cores.

By the way, this is a comment related to the accompanying paper, so this shouldn't affect the JOSS review process:
In the gravitational wave example, it would be nice to show the configuration of your run. The reason why I am asking is the RA and DEC distribution seems pretty unimodal, which is usually the case only the signal is generated with at least 3 detectors operating. There are other features in the posterior that may raise suspicion, adding the configuration you used to generate that will at least help the readers understanding whether certain features are expected or not.
