### 2022-10-09 minaskar/pocomc#17 @minaskar
Hi @kazewong and apologies for the delayed response.

The main reason that I think I need to use a "delayed" likelihood is that the Gaussian mixture or the Rosenbrock example likelihoods are very fast. Both of them are vectorised, but even if they weren't the computational communication overhead (order of 10ms) introduced by multiprocessing or MPI would dominate the total computational cost. This means that in cases in which the likelihood is cheap the parallel version will be slower than the serial. However, as the PMC algorithm (or even the vanilla SMC) is "embarrassingly parallel" scaling to many CPUs is linear assuming that the cost of a single likelihood evaluation is greater than ~10ms.  We included a short paragraph on this in the Discussion section of the accompanying paper.

Many thanks for the suggestion regarding the accompanying paper, we will certainly look into this.
