### 2022-10-09 #4634 @minaskar
@marylou-gabrie We're currently using the samples from the current annealed distribution in order to train the flow that will serve as the preconditioner for the next annealed distribution. Since the beta-spacings are small enough we found no benefit in reweighting the samples.

@dfm Thanks for the reminder and apologies for the delayed response, we will address the remaining comments during the next few days.
