+++ joss/paper.bib
@@ -1,8 +1,12 @@
-@article{karamanis2022pmc,
+@article{karamanis2022accelerating,
   title={Accelerating astronomical and cosmological inference with Preconditioned Monte Carlo},
   author={Karamanis, Minas and Beutler, Florian and Peacock, John A and Nabergoj, David, and Seljak, Uro\v{s}},
-  journal={in prep},
-  year={2022}
+  journal={Monthly Notices of the Royal Astronomical Society},
+  volume={516},
+  number={2},
+  pages={1644--1653},
+  year={2022},
+  publisher={Oxford University Press}
 }
 
 @article{trotta2017bayesian,
@@ -71,4 +75,58 @@
   author={Vretinaris, George and Vretinaris, Stamatis and Mermigkas, Christos and Karamanis, Minas and Stergioulas, Nikolaos},
   journal={in prep},
   year={2022}
+}
+
+@inproceedings{jia2020normalizing,
+  title={Normalizing constant estimation with Gaussianized bridge sampling},
+  author={Jia, He and Seljak, Uros},
+  booktitle={Symposium on Advances in Approximate Bayesian Inference},
+  pages={1--14},
+  year={2020},
+  organization={PMLR}
+}
+
+@article{hoffman2019neutra,
+  title={Neutra-lizing bad geometry in hamiltonian monte carlo using neural transport},
+  author={Hoffman, Matthew and Sountsov, Pavel and Dillon, Joshua V and Langmore, Ian and Tran, Dustin and Vasudevan, Srinivas},
+  journal={arXiv preprint arXiv:1903.03704},
+  year={2019}
+}
+
+@article{gabrie2021efficient,
+  title={Efficient bayesian sampling using normalizing flows to assist markov chain monte carlo methods},
+  author={Gabri{\'e}, Marylou and Rotskoff, Grant M and Vanden-Eijnden, Eric},
+  journal={arXiv preprint arXiv:2107.08001},
+  year={2021}
+}
+
+@inproceedings{brofos2022adaptation,
+  title={Adaptation of the Independent Metropolis-Hastings Sampler with Normalizing Flow Proposals},
+  author={Brofos, James and Gabri{\'e}, Marylou and Brubaker, Marcus A and Lederman, Roy R},
+  booktitle={International Conference on Artificial Intelligence and Statistics},
+  pages={5949--5986},
+  year={2022},
+  organization={PMLR}
+}
+
+@article{gabrie2022adaptive,
+  title={Adaptive Monte Carlo augmented with normalizing flows},
+  author={Gabri{\'e}, Marylou and Rotskoff, Grant M and Vanden-Eijnden, Eric},
+  journal={Proceedings of the National Academy of Sciences},
+  volume={119},
+  number={10},
+  pages={e2109420119},
+  year={2022},
+  publisher={National Acad Sciences}
+}
+
+@article{moss2020accelerated,
+  title={Accelerated Bayesian inference using deep learning},
+  author={Moss, Adam},
+  journal={Monthly Notices of the Royal Astronomical Society},
+  volume={496},
+  number={1},
+  pages={328--338},
+  year={2020},
+  publisher={Oxford University Press}
 }
