### 2022-10-17 fe44dc0
CC -> gamma
