+++ README.md
@@ -77,10 +77,14 @@ Please cite the following papers if you found this code useful in your research:
 
 ```bash
 @article{karamanis2022accelerating,
-    title={Accelerating astronomical and cosmological inference with Preconditioned Monte Carlo},
-    author={Karamanis, Minas and Beutler, Florian and Peacock, John A and Nabergoj, David and Seljak, Uros},
-    journal={arXiv preprint arXiv:2207.05652},
-    year={2022}
+    title={Accelerating astronomical and cosmological inference with preconditioned Monte Carlo},
+    author={Karamanis, Minas and Beutler, Florian and Peacock, John A and Nabergoj, David and Seljak, Uro{\v{s}}},
+    journal={Monthly Notices of the Royal Astronomical Society},
+    volume={516},
+    number={2},
+    pages={1644--1653},
+    year={2022},
+    publisher={Oxford University Press}
 }
 
 @article{karamanis2022pocomc,
