### 2022-11-07 #4634 @dfm
@kazewong — Great! In that case can you update [your checklist above](https://github.com/openjournals/joss-reviews/issues/4634#issuecomment-1201257586)? I'm still seeing some boxes unchecked. Thanks!
