+++ joss/paper.bib
@@ -1,132 +1,201 @@
 @article{karamanis2022accelerating,
-  title={Accelerating astronomical and cosmological inference with Preconditioned Monte Carlo},
-  author={Karamanis, Minas and Beutler, Florian and Peacock, John A and Nabergoj, David, and Seljak, Uro\v{s}},
-  journal={Monthly Notices of the Royal Astronomical Society},
-  volume={516},
-  number={2},
-  pages={1644--1653},
-  year={2022},
-  publisher={Oxford University Press}
+ author = {Karamanis, Minas and Beutler, Florian and Peacock, John A and Nabergoj, David and Seljak, Uro\v{s}},
+ title = {Accelerating astronomical and cosmological inference with preconditioned {Monte} {Carlo}},
+ journal = {Mon. Not. R. Astron Soc.},
+ volume = {516},
+ number = {2},
+ pages = {1644--1653},
+ year = {2022},
+ publisher = {Oxford University Press (OUP)},
+ doi = {10.1093/mnras/stac2272},
+ source = {Crossref},
+ url = {https://doi.org/10.1093/mnras/stac2272},
+ issn = {0035-8711, 1365-2966},
+ month = aug,
 }
 
 @article{trotta2017bayesian,
-  title={Bayesian methods in cosmology},
-  author={Trotta, Roberto},
-  journal={arXiv preprint arXiv:1701.01467},
-  year={2017}
+ author = {Trotta, Roberto},
+ title = {{Bayesian} Methods in Cosmology},
+ journal = {arXiv e-prints},
+ keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Statistics - Methodology},
+ year = {2017},
+ month = jan,
+ eid = {arXiv:1701.01467},
+ pages = {arXiv:1701.01467},
+ archiveprefix = {arXiv},
+ eprint = {1701.01467},
+ primaryclass = {astro-ph.CO},
+ adsurl = {https://ui.adsabs.harvard.edu/abs/2017arXiv170101467T},
+ adsnote = {Provided by the SAO/NASA Astrophysics Data System},
 }
 
 @article{sharma2017markov,
-  title={Markov chain Monte Carlo methods for Bayesian data analysis in astronomy},
-  author={Sharma, Sanjib},
-  journal={arXiv preprint arXiv:1706.01629},
-  year={2017}
+ author = {Sharma, Sanjib},
+ title = {{Markov} Chain {Monte} {Carlo} Methods for {Bayesian} Data Analysis in Astronomy},
+ journal = {Annu. Rev. Astron. Astr.},
+ year = {2017},
+ doi = {10.1146/annurev-astro-082214-122339},
+ number = {1},
+ source = {Crossref},
+ url = {https://doi.org/10.1146/annurev-astro-082214-122339},
+ volume = {55},
+ publisher = {Annual Reviews},
+ issn = {0066-4146, 1545-4282},
+ pages = {213--259},
+ month = aug,
 }
 
 @article{skilling2006nested,
-  title={Nested sampling for general Bayesian computation},
-  author={Skilling, John},
-  journal={Bayesian analysis},
-  volume={1},
-  number={4},
-  pages={833--859},
-  year={2006},
-  publisher={International Society for Bayesian Analysis}
+ author = {Skilling, John},
+ title = {Nested sampling for general {Bayesian} computation},
+ journal = {Bayesian Anal.},
+ volume = {1},
+ number = {4},
+ pages = {833--859},
+ year = {2006},
+ publisher = {Institute of Mathematical Statistics},
+ doi = {10.1214/06-ba127},
+ source = {Crossref},
[…]
