+++ joss/paper.md
@@ -17,7 +17,7 @@ authors:
   - name: John A. Peacock
     orcid: 0000-0002-1168-8299
     affiliation: 1
-  - name: Uros Seljak
+  - name: Uro\v{s} Seljak
     orcid: 0000-0003-2262-356X
     affiliation: 3
 
@@ -39,7 +39,7 @@ bibliography: paper.bib
 cosmology. The code is designed to sample efficiently from posterior distributions
 with non-trivial geometry, including strong multimodality and non-linearity. To this end,
 `pocoMC` relies on the Preconditioned Monte Carlo algorithm which utilises a Normalising
-Flow in order to decorrelate the parameters of the posterior. It facilitates both tasks of
+Flow to decorrelate the parameters of the posterior. It facilitates both tasks of
 parameter estimation and model comparison, focusing especially on computationally expensive
 applications. It allows fitting arbitrary models defined as a log-likelihood function and a
 log-prior probability density function in Python. Compared to popular alternatives (e.g.
@@ -49,12 +49,12 @@ manifests linear scaling.
 
 # Statement of need
 
-Over the past few decades the volume of astronomical and cosmological data has 
+Over the past few decades, the volume of astronomical and cosmological data has 
 increased substantially. At the same time, theoretical and phenomenological models
 in these fields have grown even more complex. As a response to that, a number of methods
 aiming at efficient Bayesian computation have been developed with the sole task of
 comparing those models to the available data [@trotta2017bayesian; @sharma2017markov]. 
-In the Bayesian context, scientific inference proceeds though the use of Bayes' theorem:
+In the Bayesian context, scientific inference proceeds through the use of Bayes' theorem:
 \begin{equation}\label{eq:bayes}
 \mathcal{P}(\theta) = \frac{\mathcal{L}(\theta)\pi(\theta)}{\mathcal{Z}}
 \end{equation}
@@ -70,13 +70,13 @@ evidences yields the Bayes factor which is the cornerstone of Bayesian model com
 
 Markov chain Monte Carlo (MCMC) has been established as the standard tool for 
 Bayesian computation in astronomy and cosmology, either as a standalone algorithm
-or as part of another method (e.g. nested sampling [@skilling2006nested]). However, 
-as MCMC relies on the local exploration of the posterior, the presense of non-linear
+or as part of another method [e.g., nested sampling, @skilling2006nested]. However, 
+as MCMC relies on the local exploration of the posterior, the presence of a non-linear
 correlation between parameters and multimodality can at best hinder its performance
-and at worst violate its theoretical guarantees of convergence (i.e. ergodicity). Usually
+and at worst violate its theoretical guarantees of convergence (i.e. ergodicity). Usually,
 those challenges are partially addressed by reparameterising the model using a common
 change-of-variables parameter transformation. However, guessing the right kind of
-reparameterisation a priori is not trivial as it often requires a deep knowledge of
+reparameterisation _a priori_ is not trivial as it often requires a deep knowledge of
 the physical model and its symmetries. These problems are usually complicated further by the substantial
 computational cost of evaluating astronomical and cosmological models. `pocoMC` is 
 designed to tackle exactly these kinds of difficulties by automatically reparameterising
@@ -87,8 +87,6 @@ efficiency and robustness. Compared to popular alternatives such as nested sampl
 can reduce the computational cost, and thus, the total run time of the analysis by orders of magnitude,
 in both artificial and realistic applications [@karamanis2022accelerating]. Finally, the code is well-tested
 and is currently used for research work in the field of gravitational wave parameter estimation [@vretinaris2022postmerger].
-
-![Logo of `pocoMC`.](./../logo.png)
 
 # Method
 
[…]
