+++ joss/paper.bib
@@ -130,7 +130,7 @@
 
 @article{hoffman2019neutra,
  author = {Hoffman, Matthew and Sountsov, Pavel and Dillon, Joshua V. and Langmore, Ian and Tran, Dustin and Vasudevan, Srinivas},
- title = {\{NeuTra\}-lizing Bad Geometry in {Hamiltonian} {Monte} {Carlo} Using Neural Transport},
+ title = {{NeuTra}-lizing Bad Geometry in {Hamiltonian} {Monte} {Carlo} Using Neural Transport},
  journal = {arXiv e-prints},
  keywords = {Statistics - Computation, Statistics - Machine Learning},
  year = {2019},
@@ -161,7 +161,7 @@
 
 @inproceedings{brofos2022adaptation,
  author = {Brofos, James and Gabri\'e, Marylou and Brubaker, Marcus A and Lederman, Roy R},
- title = {Adaptation of the Independent Metropolis-\{Hastings\} Sampler with Normalizing Flow Proposals},
+ title = {Adaptation of the Independent {Metropolis}-{Hastings} Sampler with Normalizing Flow Proposals},
  booktitle = {International Conference on Artificial Intelligence and Statistics},
  pages = {5949--5986},
  year = {2022},
