### 2022-11-07 #4634 @dfm
@minaskar  don't worry about the DOIs: I'll have some larger edits to the bibliography that will include them!
