### 2022-11-09 #4634 @dfm
@minaskar — Great! Can you make an updated archive using a clean checkout or directly from this zip file: https://github.com/minaskar/pocomc/archive/refs/tags/0.2.3.zip ? The version at 10.5281/zenodo.7306240 contains a lot of temporary files and build products. Thanks!
