+++ docs/build/html/_sources/install.rst.txt
@@ -1,37 +1,0 @@
-.. _install:
-
-Installation
-============
-
-Dependencies
-------------
-
-**pocoMC** depends on ``numpy``, ``torch``, ``tqdm``,  ``scipy``, ``corner`` and ``matplotlib``.
-
-Using pip
----------
-
-The easiest way to install the most recent stable version of ``pocomc`` is
-with `pip <http://www.pip-installer.org/>`_:
-
-.. code-block:: bash
-
-    pip install pocomc
-
-
-From source
------------
-
-Alternatively, you can get the source by cloning `the git
-repository <https://github.com/minaskar/pocomc>`_:
-
-.. code-block:: bash
-
-    git clone https://github.com/minaskar/pocomc.git
-
-Once you've downloaded the source, you can navigate into the root source
-directory and run:
-
-.. code-block:: bash
-
-    pip install .
