+++ docs/build/html/_sources/api.rst.txt
@@ -1,16 +1,0 @@
-=============
-API Reference
-=============
-
-``pocoMC`` consists mainly of the following parts:
-
-.. toctree::
-    :maxdepth: 1
-
-    api/sampler
-    api/flow
-    api/plotting
-    api/tools
-    api/mcmc
-    api/scaler
-
