### 2022-11-09 #4634 @dfm
@kazewong, @marylou-gabrie — many thanks for your reviews here! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨

@minaskar — your paper is now accepted and published in JOSS :zap::rocket::boom:
